transfered code from main to button and added includes to main
