+++ button.h
@@ -12,7 +12,7 @@
 extern "C" {
 #endif
 
-
+void init_button();
 
 
 #ifdef	__cplusplus
